ai-service/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
pdfplumber==0.11.7
pytesseract==0.3.13
Pillow==11.3.0
python-multipart==0.0.21
# RAG System Dependencies
faiss-cpu>=1.9.0
numpy>=1.24.0
langchain>=0.1.0
langchain-community>=0.0.10

# Free API Providers
groq>=0.4.0
sentence-transformers>=2.2.0
torch>=2.0.0